backups/backup_20260205_0150/requirements.txt
streamlit
pandas
numpy
plotly
yfinance
borsapy
python-dotenv
scipy
supabase
google-generativeai
beautifulsoup4
httpx
lxml
openpyxl
pydantic
